Atualizando codigo

+++ src/main/java/com/generation/lojagames/controller/CategoriaController.java
@@ -2,8 +2,12 @@ package com.generation.lojagames.controller;
 
 import com.generation.lojagames.model.CategoriaGame;
 import com.generation.lojagames.repository.CategoriaGameRepository;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.Optional;
@@ -25,20 +29,34 @@ public class CategoriaController {
         return categoriaGameRepository.findById(id);
     }
 
+    @GetMapping("/tipoGame/{tipoGame}")
+    public ResponseEntity<List<CategoriaGame>> getByTitle(@PathVariable String categoria) {
+        return ResponseEntity.ok(categoriaGameRepository
+                .findAllByTipoGameContainingIgnoreCase(categoria));
+    }
+
     @PostMapping
-    public CategoriaGame createCategoriaGame(@RequestBody CategoriaGame categoriaGame) {
-        return categoriaGameRepository.save(categoriaGame);
+    public ResponseEntity<CategoriaGame> post(@Valid @RequestBody CategoriaGame categoriaGame) {
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(categoriaGameRepository.save(categoriaGame));
     }
 
-    @PutMapping("/{id}")
-    public CategoriaGame updateCategoriaGame(@PathVariable Long id, @RequestBody CategoriaGame categoriaGameDetails) {
-        CategoriaGame categoriaGame = categoriaGameRepository.findById(id).orElseThrow();
-        categoriaGame.setTipoGame(categoriaGameDetails.getTipoGame());
-        return categoriaGameRepository.save(categoriaGame);
+    @PutMapping
+    public ResponseEntity<CategoriaGame> put(@Valid @RequestBody CategoriaGame categoriaGame){
+        return categoriaGameRepository.findById(categoriaGame.getId())
+                .map(resposta -> ResponseEntity.status(HttpStatus.CREATED)
+                        .body(categoriaGameRepository.save(categoriaGame)))
+                .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{id}")
-    public void deletCategoriaGame(@PathVariable Long id) {
+    public void delete(@PathVariable Long id) {
+        Optional<CategoriaGame> categoriaGame = categoriaGameRepository.findById(id);
+
+        if(categoriaGame.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
         categoriaGameRepository.deleteById(id);
     }
 }
